[utils] Beautify code

+++ deep_utils/utils/dir_utils/main.py
@@ -9,13 +9,13 @@ from deep_utils.utils.os_utils.os_path import split_extension
 
 
 def transfer_directory_items(
-    in_dir,
-    out_dir,
-    transfer_list,
-    mode="cp",
-    remove_out_dir=False,
-    skip_transfer=False,
-    remove_in_dir=False,
+        in_dir,
+        out_dir,
+        transfer_list,
+        mode="cp",
+        remove_out_dir=False,
+        skip_transfer=False,
+        remove_in_dir=False,
 ):
     """
 
@@ -63,14 +63,14 @@ def transfer_directory_items(
 
 
 def dir_train_test_split(
-    in_dir,
-    train_dir="./train",
-    val_dir="./val",
-    test_size=0.1,
-    mode="cp",
-    remove_out_dir=False,
-    skip_transfer=False,
-    remove_in_dir=False,
+        in_dir,
+        train_dir="./train",
+        val_dir="./val",
+        test_size=0.1,
+        mode="cp",
+        remove_out_dir=False,
+        skip_transfer=False,
+        remove_in_dir=False,
 ):
     from sklearn.model_selection import train_test_split
 
@@ -98,13 +98,13 @@ def dir_train_test_split(
 
 
 def split_dir_of_dir(
-    in_dir,
-    train_dir="./train",
-    val_dir="./val",
-    test_size=0.1,
-    mode="cp",
-    remove_out_dir=False,
-    remove_in_dir=False,
+        in_dir,
+        train_dir="./train",
+        val_dir="./val",
+        test_size=0.1,
+        mode="cp",
+        remove_out_dir=False,
+        remove_in_dir=False,
 ):
     """
 
@@ -149,17 +149,17 @@ def split_dir_of_dir(
 
 
 def split_xy_dir(
-    x_in_dir,
-    y_in_dir,
-    x_train_dir="train/samples",
-    y_train_dir="train/targets",
-    x_val_dir="val/samples",
-    y_val_dir="val/targets",
-    mode="cp",
-    val_size=0.1,
-    skip_transfer=False,
-    remove_out_dir=False,
-    remove_in_dir=False,
+        x_in_dir,
+        y_in_dir,
+        x_train_dir="train/samples",
+        y_train_dir="train/targets",
+        x_val_dir="val/samples",
+        y_val_dir="val/targets",
+        mode="cp",
+        val_size=0.1,
+        skip_transfer=False,
+        remove_out_dir=False,
+        remove_in_dir=False,
 ):
     train_names, val_names = dir_train_test_split(
         x_in_dir,
@@ -193,12 +193,12 @@ def split_xy_dir(
 
 
 def crawl_directory_dataset(
-    dir_: str,
-    ext_filter: list = None,
-    map_labels=False,
-    label_map_dict: dict = None,
-    logger=None,
-    verbose=1,
+        dir_: str,
+        ext_filter: list = None,
+        map_labels=False,
+        label_map_dict: dict = None,
+        logger=None,
+        verbose=1,
 ) -> Union[Tuple[List[str], List[int]], Tuple[List[str], List[int], Dict]]:
     """
     crawls a directory of classes and returns the full path of the items paths and their class names
@@ -252,15 +252,23 @@ def crawl_directory_dataset(
         return x, y
 
 
-def remove_create(dir_, logger=None, verbose=1):
+def remove_create(dir_, remove=True, logger=None, verbose=1):
+    """
+    Removes and creates the input directory!
+    :param dir_:
+    :param remove: whether to remove the directory or not
+    :param logger:
+    :param verbose:
+    :return:
+    """
     import os
     import shutil
 
-    if os.path.exists(dir_):
+    if os.path.exists(dir_) and remove:
         shutil.rmtree(dir_)
-    os.makedirs(dir_)
-    log_print(
-        logger, f"Successfully reomved and created dir: {dir_}", verbose=verbose)
+    if dir_:
+        os.makedirs(dir_, exist_ok=True)
+        log_print(logger, f"Successfully removed and created dir: {dir_}", verbose=verbose)
 
 
 def mkdir_incremental(dir_path: str, base_name="exp", fix_name=None) -> Path:
@@ -297,12 +305,12 @@ def mkdir_incremental(dir_path: str, base_name="exp", fix_name=None) -> Path:
 
 
 def cp_mv_all(
-    input_dir,
-    res_dir,
-    mode="cp",
-    filter_ext: Union[list, None] = None,
-    logger=None,
-    verbose=1,
+        input_dir,
+        res_dir,
+        mode="cp",
+        filter_ext: Union[list, None] = None,
+        logger=None,
+        verbose=1,
 ):
     """
     Using shutil library all the move/copy all the files from one directory to another one!
@@ -335,19 +343,19 @@ def cp_mv_all(
 
 
 def split_segmentation_dirs(
-    in_images,
-    in_masks,
-    out_train="./train",
-    out_val="./val",
-    image_dir_name="images",
-    mask_dir_name="masks",
-    img_ext=None,
-    mask_ext=None,
-    mode="cp",
-    test_size=0.2,
-    remove_out_dir=False,
-    remove_in_dir=False,
-    skip_transfer=False,
+        in_images,
+        in_masks,
+        out_train="./train",
+        out_val="./val",
+        image_dir_name="images",
+        mask_dir_name="masks",
+        img_ext=None,
+        mask_ext=None,
+        mode="cp",
+        test_size=0.2,
+        remove_out_dir=False,
+        remove_in_dir=False,
+        skip_transfer=False,
 ):
     from sklearn.model_selection import train_test_split
